server/requirements.txt
bcrypt==2.0.0
cffi==1.6.0
click==6.6
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
jsonschema==2.5.1
MarkupSafe==0.23
pbkdf2==1.3
pycparser==2.14
pycrypto==2.6.1
requests==2.10.0
six==1.10.0
SQLAlchemy==1.0.13
Werkzeug==0.11.10
